added an update about a flaw
Let's warn users there is a potential flaw that needs to be fixed. We don't have anyone available to fix this now, but we hope that people who use this code will be willing to help out and contribute a fix. If there are no such people, we'll archive the project.